refactor: keep gateway function in correct file
- gateway_auth_actions.go should only host functions : createKubeAuthProxyInfrastructure createEnvoyFilter
- gateway_controller_actions.go only for funtions: createGatewayInfrastructure createDestinationRule syncGatewayConfigStatus
- all the rest functions should be in gateway_support.go
- gateway_util_test.go is for support testing function
