+++ internal/controller/services/gateway/gateway_controller_actions.go
@@ -30,10 +30,7 @@ import (
 	gwapiv1 "sigs.k8s.io/gateway-api/apis/v1"
 
 	"github.com/opendatahub-io/opendatahub-operator/v2/api/common"
-	infrav1 "github.com/opendatahub-io/opendatahub-operator/v2/api/infrastructure/v1"
-	serviceApi "github.com/opendatahub-io/opendatahub-operator/v2/api/services/v1alpha1"
 	"github.com/opendatahub-io/opendatahub-operator/v2/internal/controller/status"
-	"github.com/opendatahub-io/opendatahub-operator/v2/pkg/cluster"
 	cond "github.com/opendatahub-io/opendatahub-operator/v2/pkg/controller/conditions"
 	odhtypes "github.com/opendatahub-io/opendatahub-operator/v2/pkg/controller/types"
 	"github.com/opendatahub-io/opendatahub-operator/v2/pkg/resources"
@@ -41,21 +38,6 @@ import (
 
 //go:embed resources
 var gatewayResources embed.FS
-
-// This helper reduces code duplication and improves error handling consistency.
-func validateGatewayConfig(rr *odhtypes.ReconciliationRequest) (*serviceApi.GatewayConfig, error) {
-	if rr == nil {
-		return nil, errors.New("reconciliation request cannot be nil")
-	}
-	gatewayConfig, ok := rr.Instance.(*serviceApi.GatewayConfig)
-	if !ok {
-		return nil, errors.New("instance is not of type *services.GatewayConfig")
-	}
-	if gatewayConfig == nil {
-		return nil, errors.New("gatewayConfig cannot be nil")
-	}
-	return gatewayConfig, nil
-}
 
 func createGatewayInfrastructure(ctx context.Context, rr *odhtypes.ReconciliationRequest) error {
 	l := logf.FromContext(ctx).WithName("createGatewayInfrastructure")
@@ -67,7 +49,7 @@ func createGatewayInfrastructure(ctx context.Context, rr *odhtypes.Reconciliatio
 	}
 	l.V(1).Info("Creating Gateway infrastructure", "gateway", gatewayConfig.Name)
 
-	domain, err := ResolveDomain(ctx, rr.Client, gatewayConfig)
+	domain, err := resolveDomain(ctx, rr.Client, gatewayConfig)
 	if err != nil {
 		return fmt.Errorf("failed to resolve domain: %w", err)
 	}
@@ -89,111 +71,9 @@ func createGatewayInfrastructure(ctx context.Context, rr *odhtypes.Reconciliatio
 		"gateway", DefaultGatewayName,
 		"namespace", GatewayNamespace,
 		"domain", domain,
-		"certificateType", GetCertificateType(gatewayConfig))
+		"certificateType", getCertificateType(gatewayConfig))
 
 	return nil
-}
-
-func createGatewayClass(rr *odhtypes.ReconciliationRequest) error {
-	gatewayClass := &gwapiv1.GatewayClass{
-		ObjectMeta: metav1.ObjectMeta{
-			Name: GatewayClassName,
-		},
-		Spec: gwapiv1.GatewayClassSpec{
-			ControllerName: gwapiv1.GatewayController(GatewayControllerName),
-		},
-	}
-
-	return rr.AddResources(gatewayClass)
-}
-
-func handleCertificates(ctx context.Context, rr *odhtypes.ReconciliationRequest, gatewayConfig *serviceApi.GatewayConfig, domain string) (string, error) {
-	// Input validation
-	if gatewayConfig == nil {
-		return "", errors.New("gatewayConfig cannot be nil")
-	}
-	if domain == "" {
-		return "", errors.New("domain cannot be empty")
-	}
-
-	// Get certificate configuration with default fallback
-	certConfig := gatewayConfig.Spec.Certificate
-	if certConfig == nil {
-		certConfig = &infrav1.CertificateSpec{
-			Type: infrav1.OpenshiftDefaultIngress,
-		}
-	}
-
-	// Generate secret name with fallback
-	secretName := certConfig.SecretName
-	if secretName == "" {
-		secretName = fmt.Sprintf("%s-tls", gatewayConfig.Name)
-	}
-
-	switch certConfig.Type {
-	case infrav1.OpenshiftDefaultIngress:
-		return handleOpenshiftDefaultCertificate(ctx, rr, secretName)
-	case infrav1.SelfSigned:
-		return handleSelfSignedCertificate(ctx, rr, secretName, domain)
-	case infrav1.Provided:
-		return secretName, nil
-	default:
-		return "", fmt.Errorf("unsupported certificate type: %s", certConfig.Type)
-	}
-}
-
-func handleOpenshiftDefaultCertificate(ctx context.Context, rr *odhtypes.ReconciliationRequest, secretName string) (string, error) {
-	err := cluster.PropagateDefaultIngressCertificate(ctx, rr.Client, secretName, GatewayNamespace)
-	if err != nil {
-		return "", fmt.Errorf("failed to propagate default ingress certificate: %w", err)
-	}
-
-	return secretName, nil
-}
-
-func handleSelfSignedCertificate(ctx context.Context, rr *odhtypes.ReconciliationRequest, secretName string, domain string) (string, error) {
-	err := cluster.CreateSelfSignedCertificate(
-		ctx,
-		rr.Client,
-		secretName,
-		domain,
-		GatewayNamespace,
-	)
-	if err != nil {
-		return "", fmt.Errorf("failed to create self-signed certificate: %w", err)
-	}
-
-	return secretName, nil
-}
-
-func createGateway(rr *odhtypes.ReconciliationRequest, certSecretName string, domain string, gatewayName string) error {
-	// Input validation
-	if rr == nil {
-		return errors.New("reconciliation request cannot be nil")
-	}
-	if gatewayName == "" {
-		return errors.New("gateway name cannot be empty")
-	}
-	if domain == "" {
-		return errors.New("domain cannot be empty")
-	}
-
-	// Create listeners with validation
-	listeners := CreateListeners(certSecretName, domain)
-
-	// Create gateway resource with optimized structure
-	gateway := &gwapiv1.Gateway{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      gatewayName,
-			Namespace: GatewayNamespace,
-		},
-		Spec: gwapiv1.GatewaySpec{
-			GatewayClassName: gwapiv1.ObjectName(GatewayClassName),
-			Listeners:        listeners,
-		},
-	}
-
-	return rr.AddResources(gateway)
 }
 
 // createDestinationRule creates a DestinationRule for TLS configuration using embedded YAML template.
@@ -224,19 +104,6 @@ func createDestinationRule(ctx context.Context, rr *odhtypes.ReconciliationReque
 
 	l.V(1).Info("Successfully created DestinationRule configuration")
 	return rr.AddResources(&unstructuredObjects[0])
-}
-
-// This helper function optimizes the condition checking logic.
-func isGatewayReady(gateway *gwapiv1.Gateway) bool {
-	if gateway == nil {
-		return false
-	}
-	for _, condition := range gateway.Status.Conditions {
-		if condition.Type == string(gwapiv1.GatewayConditionAccepted) && condition.Status == metav1.ConditionTrue {
-			return true
-		}
-	}
-	return false
 }
 
 func syncGatewayConfigStatus(ctx context.Context, rr *odhtypes.ReconciliationRequest) error {
